[RFC] Add UnwrapFrame
Per discussion in #71, add an UnwrapFrame function
that wraps the outermost errtrace frame from an error.
As a result of this change,
it's possible to implement buildTraceTree or a variant of it
using exported functionality exclusively.

This makes it possible for users to use a different trace formatting
mechanism than the trees generated by FormatString.

Resolves #71